backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart
sqlalchemy
python-dotenv
aiohttp
agentscope
dashscope
passlib[bcrypt]
python-jose[cryptography]
